refactor(layout): extract the paragraph wrapping subsystem from TextFlowSupport
Move the inline-paragraph wrapping subsystem — the three greedy wrap loops
(plain / inline-run / markdown) and their tokenize / trim / indent helpers, plus
ParagraphIndentSpec, emptyParagraphLine, and containsMarkdownSyntax — out of the
TextFlowSupport god-class into a focused package-private ParagraphWrapping
collaborator. TextFlowSupport's prepareParagraphLayout calls in through the wrap*
entry points; autosize, list, pagination, and slicing stay behind. Pure verbatim
move, no behaviour change; TextFlowSupport drops from 1877 to 765 LOC.

Add ParagraphWrappingTest covering the greedy plain-text wrap and the markdown
probe with a deterministic fixed-width measurement.
